Conflicts:
	src/main/java/connectFour/ConnectFourControllerView.java
	src/main/java/controller/GameControllerView.java
	src/main/java/ticTacToe/TicTacToeControllerView.java

+++ src/main/java/connectFour/ConnectFourControllerView.java
@@ -12,7 +12,10 @@ import controller.AccountManager;
 import controller.GameControllerView;
 import controller.GameMenu;
 import controller.logStatType;
+
 import javafx.geometry.Pos;
+import javafx.scene.control.Menu;
+import javafx.scene.control.MenuItem;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.StackPane;
 import javafx.scene.media.AudioClip;
@@ -44,11 +47,21 @@ public class ConnectFourControllerView extends GameControllerView {
 	// Should look and play the exact same way as before.
 	private GridPane _primarypane;
 	private GameMenu menuBar;
+	// Size of 5: 0 = Overall color, 1 = Blank Circle color, 2 = Circle outline color 
+	//            3 = Player Piece color, 4 = AI Piece color 
+	private Color[] themesettings; // Used to
 	/**
 	 * Constructor for the Connect 4 controller-view sets up the board and various
 	 * listeners
 	 */
 	public ConnectFourControllerView() {
+		themesettings = new Color[5];
+		themesettings[0] = Color.BLUE;
+		themesettings[1] = Color.WHITE;
+		themesettings[2] = Color.BLACK;
+		themesettings[3] = Color.INDIANRED;
+		themesettings[4] = Color.YELLOW;
+		
 		gameName = "connect-four";
 		menuBar = GameMenu.getMenuBar(this);
 		this.setTop(menuBar);
@@ -60,6 +73,7 @@ public class ConnectFourControllerView extends GameControllerView {
 		this.setHeight(HEIGHT);
 		_primarypane.setPrefHeight(HEIGHT);
 		_primarypane.setPrefWidth(WIDTH);
+		addCustomUIOptions();
 	}
 
 	/**
@@ -97,9 +111,9 @@ public class ConnectFourControllerView extends GameControllerView {
 		placeholder = new StackPane[7][6];
 		for (int row = 0; row < 6; row++) {
 			for (int col = 0; col < 7; col++) {
-				Rectangle rect = new Rectangle(100, 100, Color.BLUE);
-				Circle circ = new Circle(43, Color.WHITE);
-				circ.setStroke(Color.BLACK);
+				Rectangle rect = new Rectangle(100, 100, themesettings[0]);
+				Circle circ = new Circle(43, themesettings[1]);
+				circ.setStroke(themesettings[2]);
 				circ.setStrokeWidth(1);
 				StackPane spane = new StackPane();
 				spane.getChildren().addAll(rect, circ);
@@ -163,7 +177,11 @@ public class ConnectFourControllerView extends GameControllerView {
 		try {
 			String fname = accountmanager.getCurUsername() + "-" + gameName + ".dat";
 			String sep = System.getProperty("file.separator");
-			String filepath = System.getProperty("user.dir") + sep + "save-data" + sep + fname;
+			String filepath = System.getProperty("user.dir") + sep + "save-data";
+			if(!new File(filepath).exists()) {
+				new File(filepath).mkdir();
+			}
+			filepath += sep + fname;
 			fos = new FileOutputStream(filepath);			
 			oos = new ObjectOutputStream(fos);
 			oos.writeObject(gameModel);
@@ -183,17 +201,22 @@ public class ConnectFourControllerView extends GameControllerView {
 	 * @return true if the load was successful, false otherwise
 	 */
 	public boolean loadSaveGame() {
+		boolean retVal = true;
 		try {
 			String fname = accountmanager.getCurUsername() + "-" + gameName + ".dat";
 			String sep = System.getProperty("file.separator");
 			String filepath = System.getProperty("user.dir") + sep + "save-data" + sep + fname;
 			File file = new File(filepath);
-			FileInputStream fis = new FileInputStream(file);
-			ObjectInputStream ois = new ObjectInputStream(fis);
-			gameModel = (ConnectFourModel) ois.readObject();
-			ois.close();
-			update(gameModel, this);
-			file.delete();
+			if(file.exists()) {
+				FileInputStream fis = new FileInputStream(file);
+				ObjectInputStream ois = new ObjectInputStream(fis);
+				gameModel = (ConnectFourModel) ois.readObject();
+				ois.close();
+				update(gameModel, this);
+				file.delete();
+			} else {
+				retVal = newGame();
+			}
 		} catch(IOException | ClassNotFoundException e) {
 			return false;
 		}
@@ -201,7 +224,7 @@ public class ConnectFourControllerView extends GameControllerView {
 		setupListeners();
 		gameModel.addObserver(this);
 		update(gameModel, this);
-		return true;
+		return retVal;
 	}
 
 	@Override
@@ -257,7 +280,7 @@ public class ConnectFourControllerView extends GameControllerView {
 		for (int row = 0; row < 6; row++) {
 			for (int col = 0; col < 7; col++) {
 				if (board[row][col] != '_') {
-					Circle circ = new Circle(43, board[row][col] == 'R' ? Color.INDIANRED : Color.YELLOW);
+					Circle circ = new Circle(43, board[row][col] == 'R' ? themesettings[3] : themesettings[4]);
 					placeholder[col][row].getChildren().add(circ);
 				}
 			}
@@ -268,18 +291,18 @@ public class ConnectFourControllerView extends GameControllerView {
 		//System.out.println("--------------");
 		
 		if (gameModel.tied()) {
-			//accountmanager.logGlobalStat(true, "Connect-Four", logStatType.TIE, 1);
-			//accountmanager.logGameStat("Connect-Four", logStatType.TIE, 0);
+			accountmanager.logGlobalStat(true, "Connect-Four", logStatType.TIE, 1);
+			accountmanager.logGameStat("Connect-Four", logStatType.TIE, 0);
 			tieSound.play();
 		} else if (gameModel.won('R') || gameModel.won('Y')) {
 			disableListeners();
 			if (gameModel.won('R')) {
-				//accountmanager.logGlobalStat(true, "Connect-Four", logStatType.WIN, 1);
-				//accountmanager.logGameStat("Connect-Four", logStatType.WIN, 0);
+				accountmanager.logGlobalStat(true, "Connect-Four", logStatType.WIN, 1);
+				accountmanager.logGameStat("Connect-Four", logStatType.WIN, 0);
 				winSound.play();
 			} else {
-				//accountmanager.logGlobalStat(true, "Connect-Four", logStatType.LOSS, 1);
-				//accountmanager.logGameStat("Connect-Four", logStatType.LOSS, 0);
+				accountmanager.logGlobalStat(true, "Connect-Four", logStatType.LOSS, 1);
+				accountmanager.logGameStat("Connect-Four", logStatType.LOSS, 0);
 				loseSound.play();
 			}
 		}
@@ -288,7 +311,88 @@ public class ConnectFourControllerView extends GameControllerView {
 	protected void updateStatistics() {
 		if (!(gameModel.won('R') || gameModel.won('Y')) && gameModel.maxMovesRemaining() > 0) {
 			accountmanager.logGlobalStat(true, "Connect-Four", logStatType.INCOMPLETE, 1);
-			//accountmanager.logGameStat("Connect-Four", logStatType.INCOMPLETE, 1);
+			accountmanager.logGameStat("Connect-Four", logStatType.INCOMPLETE, 1);
 		}
+	}
+	/**
+	 * Redraws the game GUI. Ment to be used after changing the theme.
+	 */
+	private void reDrawGameUI() {
+		placeholder = new StackPane[7][6];
+		for (int row = 0; row < 6; row++) {
+			for (int col = 0; col < 7; col++) {
+				Rectangle rect = new Rectangle(100, 100, themesettings[0]);
+				Circle circ = new Circle(43, themesettings[1]);
+				circ.setStroke(themesettings[2]);
+				circ.setStrokeWidth(1);
+				StackPane spane = new StackPane();
+				spane.getChildren().addAll(rect, circ);
+				_primarypane.add(spane, col, row);
+				placeholder[col][row] = spane;
+			}
+		}
+		
+		char[][] board = gameModel.getBoard();
+		for (int row = 0; row < 6; row++) {
+			for (int col = 0; col < 7; col++) {
+				if (board[row][col] != '_') {
+					Circle circ = new Circle(43, board[row][col] == 'R' ? themesettings[3] : themesettings[4]);
+					placeholder[col][row].getChildren().add(circ);
+				}
+			}
+		}
+	}
+	/**
+	 * 
+	 */
+	private void addCustomUIOptions() {
+		Menu thememenu = new Menu("Theme Menu");
+		MenuItem opt1 = new MenuItem("Set Default Theme");
+		opt1.setOnAction((event) -> {
+			themesettings[0] = Color.BLUE;
+			themesettings[1] = Color.WHITE;
+			themesettings[2] = Color.BLACK;
+			themesettings[3] = Color.INDIANRED;
+			themesettings[4] = Color.YELLOW;
+			reDrawGameUI();
+		});
+		MenuItem opt2 = new MenuItem("Set Alternate Theme");
+		opt2.setOnAction((event) -> {
+			themesettings[0] = Color.DARKRED;
+			themesettings[1] = Color.GREEN;
+			themesettings[2] = Color.RED;
+			themesettings[3] = Color.BLACK;
+			themesettings[4] = Color.WHITE;
+			reDrawGameUI();
+		});
+		MenuItem opt3 = new MenuItem("Set America Theme");
+		opt3.setOnAction((event) -> {
+			themesettings[0] = Color.AQUA;
+			themesettings[1] = Color.WHITE;
+			themesettings[2] = Color.GREY;
+			themesettings[3] = Color.RED;
+			themesettings[4] = Color.BLUE;
+			reDrawGameUI();
+		});
+		MenuItem opt4 = new MenuItem("Set MLG PRO Theme");
+		opt4.setOnAction((event) -> {
+			themesettings[0] = Color.BLACK;
+			themesettings[1] = Color.GREY;
+			themesettings[2] = Color.WHITE;
+			themesettings[3] = Color.GREEN;
+			themesettings[4] = Color.RED;
+			reDrawGameUI();
+		});
+		MenuItem opt5 = new MenuItem("Set Greyscale Theme");
+		opt5.setOnAction((event) -> {
+			themesettings[0] = Color.WHITE;
+			themesettings[1] = Color.WHITE;
+			themesettings[2] = Color.WHITE;
+			themesettings[3] = Color.BLACK;
+			themesettings[4] = Color.GREY;
+			reDrawGameUI();
+		});
+		thememenu.getItems().addAll(opt1,opt2,opt3,opt4,opt5);
+		menuBar.getMenus().add(thememenu);
 	}
 }
