+++ src/main/java/ticTacToe/TicTacToeControllerView.java
@@ -9,6 +9,8 @@ import java.io.ObjectOutputStream;
 import java.util.ArrayList;
 import java.util.Observable;
 
+import Gamejam.Gamejam;
+import connectFour.ConnectFourModel;
 import controller.AccountManager;
 import controller.GameControllerView;
 import controller.GameMenu;
@@ -257,40 +259,26 @@ public class TicTacToeControllerView extends GameControllerView {
 			}
 		}
 		
-		System.out.println("----------");
-		System.out.println(gameModel.toString());
-		System.out.println("----------");
+		Gamejam.DPrint("----------");
+		Gamejam.DPrint(gameModel.toString());
+		Gamejam.DPrint("----------");
 		
 		if (gameModel.tied()) {
-			//accountmanager.logGlobalStat(true, "Tic-Tac-Toe", logStatType.TIE, 1);
-			//accountmanager.logGameStat("Tic-Tac-Toe", logStatType.TIE, 0);
+			accountmanager.logGlobalStat(true, "Tic-Tac-Toe", logStatType.TIE, 1);
+			accountmanager.logGameStat("Tic-Tac-Toe", logStatType.TIE, 0);
 			tieSound.play();
 		} else if (gameModel.won('X') || gameModel.won('O')) {
 			if (gameModel.won('X')) {
-				//accountmanager.logGlobalStat(true, "Tic-Tac-Toe", logStatType.WIN, 1);
-				//accountmanager.logGameStat("Tic-Tac-Toe",  logStatType.WIN, 1);
+				accountmanager.logGlobalStat(true, "Tic-Tac-Toe", logStatType.WIN, 1);
+				accountmanager.logGameStat("Tic-Tac-Toe",  logStatType.WIN, 1);
 				winSound.play();
 			} else {
-				//accountmanager.logGlobalStat(true, "Tic-Tac-Toe", logStatType.LOSS, 1);
-				//accountmanager.logGameStat("Tic-Tac-Toe", logStatType.LOSS, 1);
+				accountmanager.logGlobalStat(true, "Tic-Tac-Toe", logStatType.LOSS, 1);
+				accountmanager.logGameStat("Tic-Tac-Toe", logStatType.LOSS, 1);
 				loseSound.play();
 			}
-
-			/*
-			 * *****Old code to figure out a stroke for the winning row/col/diagonal****
-			 * String winningDirection = gameModel.getWinningDirection(); Point[]
-			 * winningSquares = gameModel.getWinningSquares(winningDirection);
-			 * gc.setStroke(Color.BLUE);
-			 * 
-			 * switch (winningDirection) { case "horizontal": gc.strokeLine(0,
-			 * winningSquares[0].x*200+100, 600, winningSquares[2].x*200+100); break; case
-			 * "vertical": gc.strokeLine(winningSquares[0].y*200+100, 0,
-			 * winningSquares[2].y*200+100, 600); break; case "diagonal": if
-			 * (winningSquares[0].x == 0) { gc.strokeLine(0, 0, 600, 600); } else {
-			 * gc.strokeLine(0, 600, 600, 0); } break; } gc.setStroke(Color.BLACK);
-			 */
 		}
-		System.out.println("current directory = " + System.getProperty("user.dir"));
+		Gamejam.DPrint("current directory = " + System.getProperty("user.dir"));
 	}
 
 	@Override
@@ -304,12 +292,16 @@ public class TicTacToeControllerView extends GameControllerView {
 		try {
 			String fname = accountmanager.getCurUsername() + "-" + gameName + ".dat";
 			String sep = System.getProperty("file.separator");
-			String filepath = System.getProperty("user.dir") + sep + "save-data" + sep + fname;
+			String filepath = System.getProperty("user.dir") + sep + "save-data";
+			if(!new File(filepath).exists()) {
+				new File(filepath).mkdir();
+			}
+			filepath += sep + fname;
 			fos = new FileOutputStream(filepath);			
 			oos = new ObjectOutputStream(fos);
 			oos.writeObject(gameModel);
 			oos.close();
-			System.out.println("Saved the game");
+			Gamejam.DPrint("Saved the game");
 		} catch (IOException e) {
 			e.printStackTrace();
 			return false;
@@ -328,17 +320,22 @@ public class TicTacToeControllerView extends GameControllerView {
 	public boolean loadSaveGame() {
 		FileInputStream fis;
 		ObjectInputStream ois;
+		boolean retVal = true;
 		try {
 			String fname = accountmanager.getCurUsername() + "-" + gameName + ".dat";
 			String sep = System.getProperty("file.separator");
 			String filepath = System.getProperty("user.dir") + sep + "save-data" + sep + fname;
 			File file = new File(filepath);
-			fis = new FileInputStream(file);
-			ois = new ObjectInputStream(fis);
-			gameModel = (TicTacToeModel) ois.readObject();
-			ois.close();
-			file.delete();
-			System.out.println("Succesfully loaded");
+			if(file.exists()) {
+				fis = new FileInputStream(file);
+				ois = new ObjectInputStream(fis);
+				gameModel = (TicTacToeModel) ois.readObject();
+				ois.close();
+				update(gameModel, this);
+				file.delete();
+			} else {
+				retVal = newGame();
+			}
 		} catch(IOException | ClassNotFoundException e) {
 			e.printStackTrace();
 			return false;
@@ -395,7 +392,7 @@ public class TicTacToeControllerView extends GameControllerView {
 	protected void updateStatistics() {
 		if (!(gameModel.won('X') || gameModel.won('O')) && gameModel.maxMovesRemaining() > 0) {
 			accountmanager.logGlobalStat(true, "Tic-Tac-Toe", logStatType.INCOMPLETE, 1);
-			//accountmanager.logGameStat("Tic-Tac-Toe", logStatType.INCOMPLETE, 1);
+			accountmanager.logGameStat("Tic-Tac-Toe", logStatType.INCOMPLETE, 1);
 		}
 	}
 }
